fix(docs): correct devices section to ship-realistic lineup
Removes stub pages for devices that don't exist yet (Tricorder, Petraeus,
MycoTenna, Mushroom 2, FCI Probes as separate entry) and adds the FCI
entry to pair with the existing FCI Firmware doc.

Final devices lineup (10):
- Mushroom 1, Hyphae 1, MycoNode, ALARM, MycoBrain, Agaric, SporeBase (stable)
- Psathyrella (draft — partial build live)
- FCI (frontier — multi-year hardware research program)
- FCI Firmware (stable — links to existing /docs/fci-firmware)

Diff is limited to app/docs/**, lib/docs-catalog.ts. No device pages,
no NatureOS, no FUSARIUM, no footer touched.
