+++ lib/docs-catalog.ts
@@ -282,100 +282,74 @@ export const DOCS_CATALOG: DocSection[] = [
         title: "Mushroom 1",
         description: "Edge biosignal collector.",
         href: "/docs/devices/mushroom-1",
-        status: "coming-soon",
+        status: "stable",
         sources: placeholderSources("devices-mushroom-1"),
       },
       {
         title: "Hyphae 1",
         description: "Distributed soil and substrate sensor network.",
         href: "/docs/devices/hyphae-1",
-        status: "coming-soon",
+        status: "stable",
         sources: placeholderSources("devices-hyphae-1"),
       },
       {
         title: "MycoNode",
         description: "Mesh networking node for fungal-data telemetry.",
         href: "/docs/devices/myconode",
-        status: "coming-soon",
+        status: "stable",
         sources: placeholderSources("devices-myconode"),
       },
       {
         title: "ALARM",
         description: "Biological detection and early-warning sensor with TinyML inference on-device.",
         href: "/docs/devices/alarm",
-        status: "coming-soon",
+        status: "stable",
         sources: placeholderSources("devices-alarm"),
       },
       {
         title: "MycoBrain",
         description: "Edge compute appliance running MYCA workloads on-device.",
         href: "/docs/devices/mycobrain",
-        status: "coming-soon",
+        status: "stable",
         sources: placeholderSources("devices-mycobrain"),
       },
       {
         title: "Agaric",
         description: "Compact field probe with audio and bioelectric sensing.",
         href: "/docs/devices/agaric",
-        status: "coming-soon",
+        status: "stable",
         sources: placeholderSources("devices-agaric"),
-      },
-      {
-        title: "Psathyrella",
-        description: "Subsurface and aquatic probe variant.",
-        href: "/docs/devices/psathyrella",
-        status: "coming-soon",
-        sources: placeholderSources("devices-psathyrella"),
       },
       {
         title: "SporeBase",
         description: "Stationary lab and field base station for spore capture and analysis.",
         href: "/docs/devices/sporebase",
-        status: "coming-soon",
+        status: "stable",
         sources: placeholderSources("devices-sporebase"),
       },
       {
-        title: "FCI Probes",
+        title: "Psathyrella",
         description:
-          "Fungal Computing Interface probes (research grade). Electrically active substrates, not language-capable systems.",
-        href: "/docs/devices/fci-probes",
+          "In development — partial build live; documentation forthcoming.",
+        href: "/docs/devices/psathyrella",
+        status: "draft",
+        sources: placeholderSources("devices-psathyrella"),
+      },
+      {
+        title: "FCI",
+        description:
+          "Fungal Computing Interface — multi-year research program building the hardware that, together with the FCI Firmware, enables fungal compute.",
+        href: "/docs/devices/fci",
         status: "frontier",
-        sources: placeholderSources("devices-fci-probes"),
+        sources: placeholderSources("devices-fci"),
       },
       {
         title: "FCI Firmware",
-        description: "Firmware reference for the FCI probe family — pin maps, telemetry, OTA flow.",
+        description:
+          "The firmware that bridges FCI hardware to MYCA workloads — the piece we are actively making work alongside FCI to enable fungal compute.",
         href: "/docs/fci-firmware",
         status: "stable",
         sources: [{ label: "Internal", href: "/docs/fci-firmware", kind: "internal" }],
-      },
-      {
-        title: "Tricorder",
-        description: "Handheld multi-sensor unit for field surveys.",
-        href: "/docs/devices/tricorder",
-        status: "coming-soon",
-        sources: placeholderSources("devices-tricorder"),
-      },
-      {
-        title: "Petraeus",
-        description: "High-power deployment platform for ruggedized outdoor use.",
-        href: "/docs/devices/petraeus",
-        status: "coming-soon",
-        sources: placeholderSources("devices-petraeus"),
-      },
-      {
-        title: "MycoTenna",
-        description: "RF and spectrum sensing array for environmental signals.",
-        href: "/docs/devices/mycotenna",
-        status: "coming-soon",
-        sources: placeholderSources("devices-mycotenna"),
-      },
-      {
-        title: "Mushroom 2",
-        description: "Next-gen consumer and professional variant of Mushroom 1.",
-        href: "/docs/devices/mushroom-2",
-        status: "coming-soon",
-        sources: placeholderSources("devices-mushroom-2"),
       },
     ],
   },
